Integrating PPML analytical solution in to the trunk.
Currently applicable to models specified in both path and matrix form, with either covariance matrix or raw data. Support for data sets with missingness in development.

For RAM models, if PPML is applicable in such a way that an analytical solution can be computed, the model will be transformed and solved.  Optimizations to models not analytically solvable are possible and in development, but not fully implemented yet.

Adds a new model option, UsePPML.  Can be set to "Yes" or "No".  mxDisablePPML(model) and mxEnablePPML(model) set this option.

PPML applicability filter has been separated off in to another function, imxCheckPPMLApplicable.  This function returns "Solve", "PartialSolve", "Split", or NA depending on how PPML is applicable to the given model.  The return value from this function can be passed to the transform function.  By default, however, the transform function will call imxCheckPPMLApplicable.

Following the analytical solution, the backend will be called for one cycle to compute Hessians and Standard Errors.

The UsePPML option is also used to track what PPML did to the model.  After the run is complete, it will have value "Solved", "PartialSolved", or "Split".

git-svn-id: http://openmx.psyc.virginia.edu/svn/trunk@1892 df83fd30-4cca-4c36-ab9d-e8b5583ccbd1