Simplify error reporting of unrecognized arguments to mxModel
+ Instead of printing exactly what is indigestible, we just deparse the
value. This makes it a little more difficult for users to identify what
went wrong when they pass gibberish into mxModel. For example:

Old behavior:

  Error: Cannot add the following item(s) into the model: 'rpf.drm()'

New behavior:

  Error: Cannot add the following item(s) into the model: '<S4 object of class structure("rpf.1dim.drm", package = "rpf")>'

+ The internal code becames simpler and easier to maintain. That is
the main motivation for this change. I need to make another change
that is vastly simplified by this change.

+ There could be a performance benefit. I did not test this, but it
is vaguely plausible. c(first, func.call[['...']]) might take a long
time to evaluate if there are 1000s of arguments.

git-svn-id: http://openmx.psyc.virginia.edu/svn/trunk@2427 df83fd30-4cca-4c36-ab9d-e8b5583ccbd1