Change how dirty matrices are indicated
The problem is that the frontend does no dependency tracking for
expectations. Expectations are assumed independent from free parameters.
This is no longer true with E-M style algorithms. To solve this problem,
I added a version to omxMatrix. The expectation can check whether the
version is changed to decide whether to recalculate. The dirty/clean
indicator is not reused here because the matrix is really clean all the
time in the sense that it does not need to be recomputed like an algebra.

git-svn-id: http://openmx.psyc.virginia.edu/svn/trunk@2742 df83fd30-4cca-4c36-ab9d-e8b5583ccbd1