Reorg E step 1/9
The purpose of this patch series is to reorder the loops in the E step
such that the unique pattern is the outermost loop instead of the
quadrature as the outermost loop. There are at least two reasons why
this is a good idea. The first reason is that this organization permits
us to skip the maximum amount of computation for missing data. Missing
data is detected in the outermost loop and the inner loops can be
skipped entirely. The second reason why this organization wins is
because most information is responses specific. By distributing threads
over responses, there is less need for expensive between thread
synchronization.

git-svn-id: http://openmx.psyc.virginia.edu/svn/trunk@2797 df83fd30-4cca-4c36-ab9d-e8b5583ccbd1