Combine calculations into the E-step (helps two-tier more)
If you want to explore the estimation options, such as comparing
speed with/without caching, this snapshot has everything working
and selectable at runtime. Since two-tier gains so much by not
caching and regular BA81 is not slower, the next change set will
remove the cache option and a bunch of code.

git-svn-id: http://openmx.psyc.virginia.edu/svn/trunk@2852 df83fd30-4cca-4c36-ab9d-e8b5583ccbd1