Fix algebra dependency tracking
I'm not sure how to write a test case, but here is the code that
was not working.

mxAlgebra(t(F%*%(solve(bigI - A))%*%t(M)),name = "expMean"),

There were free parameters in the M matrix. Apparently, t(M)
was compiled into an algebra with populateLocations. Since
the beginning of history, we always assumed that anything
with a populate location was not an algebra. I'm not sure
when this assumption changed.

git-svn-id: http://openmx.psyc.virginia.edu/svn/trunk@3463 df83fd30-4cca-4c36-ab9d-e8b5583ccbd1