Remove extra copy of # of evaluations
Date: Fri, 16 May 2014 09:22:12 +0000
From: "Brick, Timothy" <brick@mpib-berlin.mpg.de>

I'm looking back at the logs, and I'm not sure exactly why that was.
The thing that springs to mind was that the second value was originally
a placeholder intended to eventually keep a separate number for row-wise
computations.  There was a reason somebody wanted a report of row-wise
computations, but I think we found an alternative or realized we weren't
that interested in it as a benchmark, so the idea got shelved.

I can't think of any other reason this would have been done, and I
imagine it's no longer important to keep.

git-svn-id: http://openmx.psyc.virginia.edu/svn/trunk@3468 df83fd30-4cca-4c36-ab9d-e8b5583ccbd1