Fix for R 3.1
Apparently, R no longer copies everything it sends through .Call.
This is good because it means there is a potential performance
improvement. However, OpenMx was never tested in this situation.
An easy fix is to unconditionally copy MxMatrix@values when we
import the data from R without worrying about whether the matrix
is constant or not.

git-svn-id: http://openmx.psyc.virginia.edu/svn/trunk@3473 df83fd30-4cca-4c36-ab9d-e8b5583ccbd1