Smarter memory usage by GREML with incomplete data.
With this commit, GREML will only copy matrix data from an omxMatrix
when the latter is not from an algebra and needs to have rows and
columns dropped from it.  If the omxMatrix comes from an algebra, it can
be filtered directly and referenced by an Eigen Map, since it will be
recomputed to full size later as long as it's marked dirty.  If
filtering for NAs is no concern, then using an Eigen Map is a
no-brainer.