Make mxGREMLStarter() always report NA rows.
The intent was for the user to always be able to find which rows of X
and y mxGREMLStarter() trimmed out, simply by looking at the casesToDrop
slot of the returned fitfunction.  This wasn't happening when
dropNAfromV was FALSE, even though the backend was written to behave
appropriately.  Also added a test for the correct behavior.