Reduce needed Richardson steps for model ID check.
This has the benefit of speed.  It also has the risk of being
inaccurate.  If it is inaccurate enough, then the dimension of the null
space of the Jacobian may be different and then the identification check
might give false positives/negatives.  Initial tests indicate that it's
accurate enough, even with the speed up.