Safer behavior for definition variables used outside of FIML
If the data has exactly 1 row then we load the definition
variables. Otherwise they are set to NAs.

To make this work, it is also necessary to permit NAs
to appear in the square brackets notation [1,data.foo].
This previously triggered an error immediately. However,
now NAs are coerced to NULL and result in a 0x0 matrix result.
This will generally break anything that relies on
the computation of something. The net effect is the same,
an exception, but the message and where it is thrown can
be different.