More robust profile CIs
+ With fixes to NPSOL, it seems unnecessary to try more than once to
find the optimum for upper & lower bounds.

+ Detailed information about the estimates and fit used to obtain the
lower and upper bounds are returned back to R. This can be very helpful
for diagnostics. Unfortunately, it shows up in
model$output$computes[[2]] when there is no custom compute plan. When
there is a compute plan, it will show up MxComputeConfidenceInterval's
output slot.

+ Default feasibility tolerance is reduced from 1.0e-05 to 1e-2. NPSOL
never really paid attention to the feasibility tolerance so this is not
a backward incompatible change. Some reduction in precision is necessary
because otherwise SLSQP cannot converge. Some testing showed that 1e-2
was sufficiently accurate to obtain reasonable CIs.