Prevent sporadic failures of this test.
This test uses mxTryHard(), which involves random number generation.
Therefore, the RNG seed should be set to a constant in the test script,
so that the mxTryHard() behaves the same way whenever the script is run.
This commit introduces a seed constant. and gives mxTryHard() one
additional try (which NPSOL seems to need with this seed value).