Fix initial compute with def vars when threads are enabled
omxData is not copied for each thread. Hence, the omxState that was
stored in omxData was always the global omxState. This was misleading.
The bug was that the code to load fake data into definition variables
for threads never did so but loaded fake data into the global omxState.
Hence, definition variables were set to NaN instead of 1 in threads.
In some models, this can cause things to go wrong with sequence
generators (e.g., 1:data.foo).

This change removes the omxState stored in omxData since it was
misleading and repairs the initialization sequence for definition
variables.