Prevent backend from trying to resize derivatives of V more than once.
If a derivative of V came from a frontend MxAlgebra, then it gets
resized directly to drop rows and columns corresponding to missing
observations; it is then marked dirty to ensure it gets recomputed back
to its full size.  This commit prevents the backend GREML fitfunction
from attempting to resize such matrices more than once.  This bug was
not previously detected because the existing test for GREML's
missing-data-handling only provided frontend MxMatrices, not algebras,
as derivatives of V.