Add internal plumbing for optimization with profiled-out parameters
I removed the augmented lagrange version of steepest descent because it
is unused, broken, and too much work to convert to the new API. If you
have interest, please revived it.

I had to relax some test tolerances slightly for NPSOL. I am not sure
why, but I also notice that NPSOL behaves slightly more intelligently
with probing the constraints.