Officially switch license to GPL v3+
Background:
http://stackoverflow.com/questions/2982364/how-does-apache-2-0-license-differ-from-gpl

I come to statistics with a background in software engineering. During
the last 30~years, one of the primary controversies in software
pertains to how software might be owned. Some people believe that it is
ethical for corporations or individuals to own and sell licenses to
their black-box software. Others, like myself, object to this practice.
Software is not like most physical objects. If I cook spaghetti then I
have cause to be upset if somebody else unexpectedly eats it. In
contrast, the cost of making an exact copy of software is negligible.
Moreover, with software, I still retain the original even after
somebody else has copied it.

Partially due to the unusual ease of copying software, I hold the
conviction that we ought to have the freedom to run, copy, distribute,
study, change, and improve software. Public domain software is not free
in this sense because a corporation can take public domain software,
improve it, and sell it as proprietary software. Whether the original
software was open-source is irrelevant. Software that was originally
open-source may become proprietary, close-source software without
specific protection against this practice. To my knowledge, the best
way to protect software is to leverage copyright law with a
\emph{copyleft} license. Copyleft uses copyright law to require that
all modified and extended versions of a program retain the same
freedoms as offered by the original work.

Besides the usual arguments in favor of copyleft, there are particular
reasons why statistical software ought to be protected as free
software. Reproducible analyses are greatly hindered by proprietary
software. A serious problem is that the mathematics implemented by
proprietary software cannot be verified. Moreover, companies, and the
software they publish, can become unmaintained and gradually cease to
function as newer versions of relied upon software are updated and
evolved.

A more subtle disadvantage of proprietary software is that it rejects
the thirst for knowledge.  It says, ``The knowledge you want is a
secret!'' Proprietary software is antagonistic to the spirit of
education and research. Why, after all, do some proprietary software
developers offer gratis or discounted copies of their nonfree programs
to schools? They want to use schools to implant dependence on their
products, like tobacco companies distributing gratis cigarettes to
school children.

Lastly, software developers, like myself are concerned about software
licensing because software that I have contributed to or written can be
taken away from me and owned by somebody else unless that software is
protected. My professional achievements are threatened if I do not
practice due diligence with respect to software ownership. That makes
the particular software license a powerful incentive or disincentive
when I consider whether I want to invest time in contributing to or
even learning about a software program.