Remove special case for assignment of mxData to RAM models
The original intent was to add an 'M' matrix when the newly added
mxData is type='raw'. This wrecks havoc with relational SEM models
because you can have raw data with no manifest variables and the 'M'
matrix gets set at the end of MxRun, making it difficult to re-run a
model. Fortunately, it looks like nobody relies on this behavior.
mxPath from 'one' is typically what creates the 'M' matrix.