Repair GREML dV bug.
Prior to this commit, if a derivative of the V matrix was not an actual
dependency of V, then there was not necessarily any time at which it
would be recomputed after a change in free-parameter values.

This commit also sets the groundwork for a more efficient way for the
GREML backend to deal with missing data.  When rows and columns of a
derivative of V need to be removed due to missingness, if the derivative
is from an algebra, the data-matrix of its omxMatrix is downsized
directly, and it gets marked dirty so that it is subsequently recomputed
to its full size.  This is inefficient because the downsizing occurs
with every fitfunction evaluation, even though an algebra that does not
depend upon free parameters need not be recomputed during optimization.