Refactor GREML derivatives code.
I realized today that the backend code that calculates GREML fitfunction
derivatives from the derivatives of the V matrix could be reworked to
remove O(N^3) bottlenecks, and only allocate an additional 2N doubles
per thread, instead of N^2 per thread.  The new code takes advantage of
objects that are calculated when evaluating the fit, and the fact that
the fitfunction derivatives are functions of matrix traces and of
quadratic forms in the y vector.