Rewrite way to name constraint diagnostics exported to frontend.
Making the naming function a method defined for class MxComputeGradientDescent was not the right approach.  For one thing, we already have a function that names optimizer output from the backend--it's called nameOptimizerOutput().  For another, it's pointless to do the naming for MxComputeGradientDescent objects that are subordinate to an MxComputeConfidenceInterval object.