Implement "internal" warm starts for NPSOL.
The idea behind an internal warm start is as follows.  If a model's fitfunction can provide an analytic Hessian, the user employs a custom compute plan that places an MxComputeOnce step, which requests the Hessian from the fitfunction, right before the MxComputeGradientDescent step.  Then, that Hessian is automatically Cholesky-factored, and if the factoring is successful, passed to NPSOL as a warm start.