Remove horribly broken PPML NHEV code
Code to recognize particular constraints was broken by increased
strictness of enforcement of correct slot types by R 3.3.x. The
constraint was originally written as 'E2' == 0.5 * 'E1' (with quotes to
turn the variable names into strings). When I corrected this to
E2 == 0.5 * E1 then things got horribly broken because the syntax
tree is different. Instead of rewritting all the syntax tree
parsing code, it was judged expedient to remove the feature and
the test.