Filter numerical gradient for outliers
In FIML or quadrature integration, some of the row/points can obtain
exactly zero probability. To cope with this, we have
addSkippedRowPenalty and mxRobustLog which increase the -2LL by a large
amount. When we try to take the numerical gradient in an area on the
boundary between valid and invalid probability then we can obtain very
large gradients. The Euclidian norm of the gradient vector is used as
information by the optimizer. To avoid confusing the optimizer, we need
to ensure that the gradients entries are roughly the same order of
magnitude.