Nelder-Mead: repair inequality-constraint-method bug, and disable pseudoHessian when using l1p algorithm.
My design was for the user to be able to request that his/her choice of equality-constraint method also be used for
inequalities.  But, prior to this commit, that didn't happen if there were no equality constraints present in the model.

There are some good reasons not to bother with the curvature matrix and the Kelley simplex matrix when using the l1
penalty algorithm.  For one, the curvature matrix would approximate the Hessian of the penalized fitfunction, and not the
-2logL, obviating its use for statistical inference.  For another, the penalty term makes the fitfunction discontinuous.