Rampart sufficient statistic clashed with zero variance predictors
Units with identical covariance can be included in a Rampart rotation
batch. At the lowest level, the expected means are often the same for
most units resulting in identical mean structure (all zeros). However,
if there are zero variance predictors in the upper levels then the
expected means will differ among lowest level units. This is not tracked
by the usual mechanics because the mean structure is disturbed only by
the Rampart rotation. That is, prior to rotation we might have lots of
identical expected means but due to the inclusion of different upper
levels (with different zero variance def vars) within the same rotation
batch, the expected means will be disturbed. In this case, we simply
disable the sufficient statistic optimization.