Populate x0 and P0 with definition variables from the first row
This fixes some impolite behavior (sort of a bug) reported by
Charlie Driver over email on 'Thu, 11 May 2017 18:23:25'.

When a definition variable was used in the state space initial
means and/or variances (x0, P0), the fit function was always
NaN.  This happened because at the time x0 and P0 were reset
no definition variables were loaded.  Now we use the first row
of definition variables to populate into x0 and P0.