Saner bootstrap quantiles
The 'max(c(-Inf,subset' etc. code was something I wrote to define the p empirical quantile, where p is on [0,1], as the largest order statistic such that no more than p of the data was less than or equal to it.  I recently realized that's backwards--the p empirical quantile is some number such that at least p of the data is less than or equal to it.  Since empirical quantiles aren't necessarily unique, using the smallest such order statistic is a sensible way to proceed, and that's what quantile() does with type=1.