WLS: Move observed stats to as separate slot in the MxData object
NEWS: This is a potentially incompatible change because the
observed stats are accessed under data$observedStats. However,
it is necessary to better delineate the observedStats (which
are a function of the data) from the actual data. mxDataWLS
now retains the raw data in the observed slot instead of the
covariance matrix.