WLS: Deal with missing data pairwise
This doesn't match lavaan. I suspect this is because the denominators
are inconsistent. Sometimes the denominator is the number of rows
and sometimes it is the number of non-missing rows. It probably
makes more sense to divide by the number of non-missing rows, but
there probably isn't a best choice here. It's a compromise either way.