Point to new libnpsol.a for gcc-8
Allows OpenMx to be built with both NPSOL and one of the fast BLAS
libraries (OpenBLAS or Intel's MKL). This often works but can
segfault. NPSOL continues to work correctly when linked against
the default BLAS. So while not ideal, this new binary offers
more flexibility with no loss of functionality.