Rampart: Fix bug that could inappropriately rotate at levels > 1
This bug was probably hiding in the code since the initial
release of Rampart. A test was added in Oct 2018 that started
failing with the release of R 3.6 due to its new random number
generator. This failure was detected yesterday. The bug only
affected datasets with more than two levels. The comparison
between candidate rows was being short-circuited so they
were regarded as having identical covariance when this was
not the case. This resulted in the performance of invalid
rotations.