Tweak behavior of omxRecompute
This is a fairly risky change, but it passes 'make test'. Without
this change, the matrix version & cleanVersion update every
single time you call omxRecompute. With this change, the versions
don't change for no reason. The performance gain can be on the
order of 10% so it seems worthwhile to try.