GREML: prevent buffer overrun in single-thread computation of semiAnalyt derivs.
The problem was that, when running single-threaded, `crude_numeric_dV()` did
not resize the V matrix to account for missing observations; it just
recomputed V at the current parameter values.  But now, the function just
assigns, to a correctly-sized V, its elements as previously computed at the
current parameter values--and it does that no matter how many threads are in
use.  The current behavior will generally be more computationally efficient.

This is an instance where the mantra to "just keep the code simple in the
single-threaded case" led to slower and less-safe code.