Prevent data.frame from dropping to class vector
If the model has only 1 endogenous variable, the data drop to class == "vector" when subsetting with `data <- data[,useVars]`. Subsequently, the code `data <- as.matrix(data)` returns the data to tabular format - but loses the original class of the endogenous variable in the process. For example, if we start with a single variable of class "ordered", we will end up with a single variable of class "character". This is undesirable. By using the argument `, drop = FALSE` when subsetting, this problem is avoided entirely.