Create DEBUG_CXXFLAG DEBUGMX_NEWSTUFF .
The idea is that you would uncomment the appropriate line of src/Makevars.in
to define the flag `DEBUGMX_NEWSTUFF`, and thus define as 1 the backend
variable `OMX_DEBUG_NEWSTUFF`, in order to debug new code; your new
code should be conditioned on `OMX_DEBUG_NEWSTUFF`.  You are not supposed
to permanently condition any code on `OMX_DEBUG_NEWSTUFF`--once you are
finished debugging, either remove the conditioning altogether, or condition on
one of the other OMX_DEBUG* variables.