Create new mxOption, "Analytic RAM derivatives".
The new mxOption specfically governs analytic gradients and Hessians for (a
subset of) RAM models.  To actually use such analytic derivatives with an
"eligible" RAM model, both this new mxOption and the pre-existing
"Analytic gradients" mxOption must be switched on.  As of this writing, the
on-load default for the new mxOption is "No".