Make sure NPSOL's gradient makes it into the FitContext.
The idea is that the gradient-based optimizer's gradient should be saved after
optimization is complete, and if there is no MxComputeNumericDeriv step, then
the optimizer's gradient will appear in the MxModel's 'output' slot; if there
is an MxComputeNumericDeriv step, then the optimizer's gradient gets
overwritten by the "final" gradient at that time.

With this commit, the `workingGrad` of `omxNPSOL1()` gets copied to the `grad`
member of the GradientOptimizerContext, which, via its member function
`finish()`, copies it to the `gradZ` member of the FitContext.

Prior to this commit, the gradient in the MxModel's output slot would be all
`NaN`s if the MxModel were run with NPSOL and with mxOption "Calculate Hessian"
set to "No" (and with the default compute plan).