Update R/MxFactorScores.R : matrix-specfified RAM models (#373).
It's possible that ML/WML factor scoring never worked for matrix-specified RAM
models; regression scoring certainly never did.  We lack a test of
`mxFactorScores()` used with matrix-specified RAM models.

The pre-existing code made some assumptions that were true for path-specified
RAM models, but usually false for matrix-specified RAM models.  The current
code makes assumptions that are still not guaranteed to be true for
matrix-specified models, but if they are false, then the user is not following
the guidance of the man page for `mxExpectationRAM()`.

The current code should be more robust...for example, some of the old code
assumed that the 'F' matrix always has dimnames, or that the 'A', 'S', 'F',
and 'M' matrices are actually named "A", "S", "F", and "M".